+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [0.1.2 - Unreleased] - 07-June-2021
 
 - Add friendly crash messages
+- Migrate to kubectl-view-allocations library
 
 ## [0.1.1] - 04-June-2021